backend/requirements.txt
astroid==2.12.10
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pyasn1==0.4.8
pycryptodome==3.15.0
pylint==2.15.3
python-jose==3.3.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
tomlkit==0.11.4
typed-ast==1.5.4
Werkzeug==2.2.2
wrapt==1.14.1
